Reworked configuration reducer
