+++ plugin-hrm-form/src/states/ActionTypes.js
@@ -24,10 +24,6 @@ export const HANDLE_EXPAND_DETAILS_SECTION = 'HANDLE_EXPAND_DETAILS_SECTION';
 export const QUEUES_STATUS_UPDATE = 'QUEUES_STATUS_UPDATE';
 export const QUEUES_STATUS_FAILURE = 'QUEUES_STATUS_FAILURE';
 
-// Configuration State
-export const CHANGE_LANGUAGE = 'CHANGE_LANGUAGE';
-export const POPULATE_COUNSELORS = 'POPULATE_COUNSELORS';
-
 // Categories
 export const SET_CATEGORIES_GRID_VIEW = 'SET_CATEGORIES_GRID_VIEW';
 export const HANDLE_EXPAND_CATEGORY = 'HANDLE_EXPAND_CATEGORY';
